Renaming 'endpoint' to 'controller'
